2001-05-31  Dave Peticolas  <dave@krondo.com>
	* src/engine/sixtp-utils.c (string_to_gint32): use intermediate
	variable in case int != gint32.

	* src/test/test-dom-converters1.c: fix spelling

	* src/engine/sixtp-dom-generators.c (int_to_dom_tree): cast
	gint64 to long long int for %lld.


git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@4345 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/engine/sixtp-dom-generators.c
@@ -54,8 +54,8 @@ int_to_dom_tree(const char *tag, gint64 val)
 {
     gchar *text;
     xmlNodePtr result;
-    
-    text = g_strdup_printf("%lld", val);
+
+    text = g_strdup_printf("%lld", (long long int) val);
     result = text_to_dom_tree(tag, text);
     g_free(text);
     return result;
